Resolve conflicts, convert PR #829, and make pretty
